+++ CHANGES.md
@@ -15,6 +15,7 @@ Changes in 11.6
 - CSS add .has-input class to list having input (input or select fields) in ListOutput.fnc.php
 - CSS add .list-save class to Export list link in ListOutput.fnc.php
 - CSS add .list-add-row class (see `$link['add']` parameter) in ListOutput.fnc.php
+- HTML add data-list-id attribute to list in ListOutput.fnc.php
 - ExplodeDate() use strtotime() in Date.php
 - Remove MonthNWSwitch(), `__mnwswitch_num2char()` & `__mnwswitch_char2num()` functions in Date.php
 - Send wkhtmltopdf error by email in PDF.php